Cannot use template filters inside blocktrans tags
1b549d1b made product titles disappear from the product alert emails;
I assume it was tested in custom templates that were not wrapped in
`{% blocktrans %}`.

From the Django docs "To translate a template expression – say, accessing
object attributes or using template filters – you need to bind the expression
to a local variable for use within the translation block.

The templates in question are txt emails so I have used `{% autoescape off %}`
on the whole content as any escaping will be unwelcome and of no benefit. An
alternative would be `{% blocktrans title=alert.product.get_title|safe %}`

Note, I have also standardised on the use of `Product.get_title` in these
templates as that seems to be the more common than using `Product.__str__`?