Fix AutoSlugField.deconstruct()
The create_slug() method modifies the property `_populate_from` which is
part of the deconstruct() return values. This causes the Django
migration checker to always think that there are outstanding changes to
the field. This is normally not an issue, but if you want to check for
missing migrations during an unittest this will happen.

Note that this is a simple workaround for now by keeping the original
populate_obj intact and returning that.