Fix site notifications generated by send_product_alerts.
Notification subjects have a max length of 255 characters, which can easily be exceeded by long product names. Change the alert logic to provide a short notification subject and put the longer content in the notification body.

This patch also modifies the notification utils to make it obvious that the positional argument is a notification subject, not a body.