Perform only 1 query to determine if options exist.
The self.options property query performs only 1 query. The query in has_options performs 2 queries in the current worst case scenario because it first queries for product class options and next for product options. the options property retrives both in 1 query so it is cheaper to use that together with exists to determine if there are any options.