Remove links for editing/deleting related objects from RelatedMultipleWidgetWrapper.
Adding HTML markup into each option requires disabling HTML escaping on
the option, which means the option name is also not escaped. While this
is not a very high risk, these extra links add marginal extra utility
and require a fair bit of Javascript, so it seems to be better just to
remove them.